refactor: move comprehension expansion code from `core::ast` to `rules`
Move comprehension expansion code from `conjure-cp-core::ast` to `conjure-cp-rules`.

This removes a circular dependency between the ast and solver modules in
`conjure-cp-core`, by making `ast` no longer contain any code that
requires `solver`.

This is necessary for a future PR that will move the AST code to a
seperate crate than the solver code.
