+++ crates/conjure-cp-core/src/ast/comprehension.rs
@@ -1,48 +1,23 @@
 #![allow(clippy::arc_with_non_send_sync)]
 
-mod expand_ac;
+use std::{cell::RefCell, collections::BTreeSet, fmt::Display, rc::Rc, sync::atomic::AtomicBool};
 
-use std::{
-    cell::RefCell,
-    collections::{BTreeSet, HashMap},
-    fmt::Display,
-    rc::Rc,
-    sync::{
-        Arc, Mutex, RwLock,
-        atomic::{AtomicBool, Ordering},
-    },
-};
-
-use expand_ac::add_return_expression_to_generator_model;
 use itertools::Itertools as _;
 use serde::{Deserialize, Serialize};
-use tracing::warn;
 use uniplate::{Biplate, Uniplate};
 
-use crate::{
-    ast::Metadata,
-    ast::{
-        Atom, DeclarationKind,
-        serde::{HasId as _, ObjId},
-    },
-    bug,
-    context::Context,
-    into_matrix_expr, matrix_expr,
-    rule_engine::{resolve_rule_sets, rewrite_morph, rewrite_naive},
-    solver::{Solver, SolverError},
-};
+use crate::{ast::Metadata, into_matrix_expr, matrix_expr};
 
-use super::{
-    DeclarationPtr, Domain, Expression, Model, Moo, Name, Range, SubModel, SymbolTable,
-    ac_operators::ACOperatorKind,
-};
+use super::{DeclarationPtr, Domain, Expression, Moo, Name, Range, SubModel, SymbolTable};
 
-// TODO: better way to specify this?
+// TODO: move this global setting somewhere better?
 
 /// The rewriter to use for rewriting comprehensions.
 ///
 /// True for optimised, false for naive
 pub static USE_OPTIMISED_REWRITER_FOR_COMPREHENSIONS: AtomicBool = AtomicBool::new(false);
+
+// TODO: remove ComprehensionKind, replace with ACOperatorKind
 
 // TODO: do not use Names to compare variables, use DeclarationPtr and ids instead
 // see issue #930
@@ -57,14 +32,19 @@ pub enum ComprehensionKind {
     And,
     Or,
 }
+
 /// A comprehension.
 #[derive(Clone, PartialEq, Eq, Uniplate, Serialize, Deserialize, Debug)]
 #[biplate(to=SubModel)]
 #[biplate(to=Expression)]
+#[non_exhaustive]
 pub struct Comprehension {
-    return_expression_submodel: SubModel,
-    generator_submodel: SubModel,
-    induction_vars: Vec<Name>,
+    #[doc(hidden)]
+    pub return_expression_submodel: SubModel,
+    #[doc(hidden)]
+    pub generator_submodel: SubModel,
+    #[doc(hidden)]
+    pub induction_vars: Vec<Name>,
 }
 
 impl Comprehension {
@@ -80,372 +60,6 @@ impl Comprehension {
             Box::new(return_expr_domain),
             vec![Domain::Int(vec![Range::UnboundedR(1)])],
         ))
-    }
-
-    /// Expands the comprehension using Minion, returning the resulting expressions.
-    ///
-    /// This method performs simple pruning of the induction variables: an expression is returned
-    /// for each assignment to the induction variables that satisfy the static guards of the
-    /// comprehension. If the comprehension is inside an associative-commutative operation, use
-    /// [`expand_ac`] instead, as this performs further pruning of "uninteresting" return values.
-    ///
-    /// If successful, this modifies the symbol table given to add aux-variables needed inside the
-    /// expanded expressions.
-    pub fn expand_simple(self, symtab: &mut SymbolTable) -> Result<Vec<Expression>, SolverError> {
-        let minion = Solver::new(crate::solver::adaptors::Minion::new());
-        // FIXME: weave proper context through
-        let mut model = Model::new(Arc::new(RwLock::new(Context::default())));
-
-        // only branch on the induction variables.
-        model.search_order = Some(self.induction_vars.clone());
-        *model.as_submodel_mut() = self.generator_submodel.clone();
-
-        // TODO:  if expand_ac is enabled, add Better_AC_Comprehension_Expansion here.
-
-        // call rewrite here as well as in expand_ac, just to be consistent
-        let extra_rule_sets = &["Base", "Constant", "Bubble"];
-
-        let rule_sets =
-            resolve_rule_sets(crate::solver::SolverFamily::Minion, extra_rule_sets).unwrap();
-
-        let model = if USE_OPTIMISED_REWRITER_FOR_COMPREHENSIONS.load(Ordering::Relaxed) {
-            rewrite_morph(model, &rule_sets, false)
-        } else {
-            rewrite_naive(&model, &rule_sets, false, false).unwrap()
-        };
-
-        // HACK: also call the rewriter to rewrite inside the comprehension
-        //
-        // The original idea was to let the top level rewriter rewrite the return expression model
-        // and the generator model. The comprehension wouldn't be expanded until the generator
-        // model is in valid minion that can be ran, at which point the return expression model
-        // should also be in valid minion.
-        //
-        // By calling the rewriter inside the rule, we no longer wait for the generator model to be
-        // valid Minion, so we don't get the simplified return model either...
-        //
-        // We need to do this as we want to modify the generator model (add the dummy Z's) then
-        // solve and return in one go.
-        //
-        // Comprehensions need a big rewrite soon, as theres lots of sharp edges such as this in
-        // my original implementation, and I don't think we can fit our new optimisation into it.
-        // If we wanted to avoid calling the rewriter, we would need to run the first half the rule
-        // up to adding the return expr to the generator model, yield, then come back later to
-        // actually solve it?
-
-        let return_expression_submodel = self.return_expression_submodel.clone();
-        let mut return_expression_model = Model::new(Arc::new(RwLock::new(Context::default())));
-        *return_expression_model.as_submodel_mut() = return_expression_submodel;
-
-        let return_expression_model =
-            if USE_OPTIMISED_REWRITER_FOR_COMPREHENSIONS.load(Ordering::Relaxed) {
-                rewrite_morph(return_expression_model, &rule_sets, false)
-            } else {
-                rewrite_naive(&return_expression_model, &rule_sets, false, false).unwrap()
-            };
-
-        let minion = minion.load_model(model.clone())?;
-
-        let values = Arc::new(Mutex::new(Vec::new()));
-        let values_ptr = Arc::clone(&values);
-
-        tracing::debug!(model=%model,comprehension=%self,"Minion solving comprehension (simple mode)");
-        minion.solve(Box::new(move |sols| {
-            // TODO: deal with represented names if induction variables are abslits.
-            let values = &mut *values_ptr.lock().unwrap();
-            values.push(sols);
-            true
-        }))?;
-
-        let values = values.lock().unwrap().clone();
-
-        let mut return_expressions = vec![];
-
-        for value in values {
-            // convert back to an expression
-
-            let return_expression_submodel = return_expression_model.as_submodel().clone();
-            let child_symtab = return_expression_submodel.symbols().clone();
-            let return_expression = return_expression_submodel.into_single_expression();
-
-            // we only want to substitute induction variables.
-            // (definitely not machine names, as they mean something different in this scope!)
-            let value: HashMap<_, _> = value
-                .into_iter()
-                .filter(|(n, _)| self.induction_vars.contains(n))
-                .collect();
-
-            let value_ptr = Arc::new(value);
-            let value_ptr_2 = Arc::clone(&value_ptr);
-
-            // substitute in the values for the induction variables
-            let return_expression = return_expression.transform_bi(&move |x: Atom| {
-                let Atom::Reference(ref ptr) = x else {
-                    return x;
-                };
-
-                // is this referencing an induction var?
-                let Some(lit) = value_ptr_2.get(&ptr.name()) else {
-                    return x;
-                };
-
-                Atom::Literal(lit.clone())
-            });
-
-            // Copy the return expression's symbols into parent scope.
-
-            // For variables in the return expression with machine names, create new declarations
-            // for them in the parent symbol table, so that the machine names used are unique.
-            //
-            // Store the declaration translations in `machine_name_translations`.
-            // These are stored as a map of (old declaration id) -> (new declaration ptr), as
-            // declaration pointers do not implement hash.
-            //
-            let mut machine_name_translations: HashMap<ObjId, DeclarationPtr> = HashMap::new();
-
-            // Populate `machine_name_translations`
-            for (name, decl) in child_symtab.into_iter_local() {
-                // do not add givens for induction vars to the parent symbol table.
-                if value_ptr.get(&name).is_some()
-                    && matches!(&decl.kind() as &DeclarationKind, DeclarationKind::Given(_))
-                {
-                    continue;
-                }
-
-                let Name::Machine(_) = &name else {
-                    bug!(
-                        "the symbol table of the return expression of a comprehension should only contain machine names"
-                    );
-                };
-
-                let id = decl.id();
-                let new_decl = symtab.gensym(&decl.domain().unwrap());
-
-                machine_name_translations.insert(id, new_decl);
-            }
-
-            // Update references to use the new delcarations.
-            #[allow(clippy::arc_with_non_send_sync)]
-            let return_expression = return_expression.transform_bi(&move |atom: Atom| {
-                if let Atom::Reference(ref decl) = atom
-                    && let id = decl.id()
-                    && let Some(new_decl) = machine_name_translations.get(&id)
-                {
-                    Atom::Reference(new_decl.clone())
-                } else {
-                    atom
-                }
-            });
-
-            return_expressions.push(return_expression);
-        }
-
-        Ok(return_expressions)
-    }
-
-    /// Expands the comprehension using Minion, returning the resulting expressions.
-    ///
-    /// This method is only suitable for comprehensions inside an AC operator. The AC operator that
-    /// contains this comprehension should be passed into the `ac_operator` argument.
-    ///
-    /// This method performs additional pruning of "uninteresting" values, only possible when the
-    /// comprehension is inside an AC operator.
-    ///
-    /// If successful, this modifies the symbol table given to add aux-variables needed inside the
-    /// expanded expressions.
-    pub fn expand_ac(
-        self,
-        symtab: &mut SymbolTable,
-        ac_operator: ACOperatorKind,
-    ) -> Result<Vec<Expression>, SolverError> {
-        // ADD RETURN EXPRESSION TO GENERATOR MODEL AS CONSTRAINT
-        // ======================================================
-
-        // References to induction variables in the return expression point to entries in the
-        // return_expression symbol table.
-        //
-        // Change these to point to the corresponding entry in the generator symbol table instead.
-        //
-        // In the generator symbol-table, induction variables are decision variables (as we are
-        // solving for them), but in the return expression symbol table they are givens.
-        let induction_vars_2 = self.induction_vars.clone();
-        let generator_symtab_ptr = Rc::clone(self.generator_submodel.symbols_ptr_unchecked());
-        let return_expression =
-            self.clone()
-                .return_expression()
-                .transform_bi(&move |decl: DeclarationPtr| {
-                    // if this variable is an induction var...
-                    if induction_vars_2.contains(&decl.name()) {
-                        // ... use the generator symbol tables version of it
-
-                        (*generator_symtab_ptr)
-                            .borrow()
-                            .lookup_local(&decl.name())
-                            .unwrap()
-                    } else {
-                        decl
-                    }
-                });
-
-        // Replace all boolean expressions referencing non-induction variables in the return
-        // expression with dummy variables. This allows us to add it as a constraint to the
-        // generator model.
-        let generator_submodel = add_return_expression_to_generator_model(
-            self.generator_submodel.clone(),
-            return_expression,
-            &ac_operator,
-        );
-
-        // REWRITE GENERATOR MODEL AND PASS TO MINION
-        // ==========================================
-
-        let mut generator_model = Model::new(Arc::new(RwLock::new(Context::default())));
-
-        *generator_model.as_submodel_mut() = generator_submodel;
-
-        // only branch on the induction variables.
-        generator_model.search_order = Some(self.induction_vars.clone());
-
-        let extra_rule_sets = &[
-            "Base",
-            "Constant",
-            "Bubble",
-            "Better_AC_Comprehension_Expansion",
-        ];
-
-        let rule_sets =
-            resolve_rule_sets(crate::solver::SolverFamily::Minion, extra_rule_sets).unwrap();
-
-        let generator_model = if USE_OPTIMISED_REWRITER_FOR_COMPREHENSIONS.load(Ordering::Relaxed) {
-            rewrite_morph(generator_model, &rule_sets, false)
-        } else {
-            rewrite_naive(&generator_model, &rule_sets, false, false).unwrap()
-        };
-
-        let minion = Solver::new(crate::solver::adaptors::Minion::new());
-        let minion = minion.load_model(generator_model.clone());
-
-        let minion = match minion {
-            Err(e) => {
-                warn!(why=%e,model=%generator_model,"Loading generator model failed, failing expand_ac rule");
-                return Err(e);
-            }
-            Ok(minion) => minion,
-        };
-
-        // REWRITE RETURN EXPRESSION
-        // =========================
-
-        let return_expression_submodel = self.return_expression_submodel.clone();
-        let mut return_expression_model = Model::new(Arc::new(RwLock::new(Context::default())));
-        *return_expression_model.as_submodel_mut() = return_expression_submodel;
-
-        let return_expression_model =
-            if USE_OPTIMISED_REWRITER_FOR_COMPREHENSIONS.load(Ordering::Relaxed) {
-                rewrite_morph(return_expression_model, &rule_sets, false)
-            } else {
-                rewrite_naive(&return_expression_model, &rule_sets, false, false).unwrap()
-            };
-
-        let values = Arc::new(Mutex::new(Vec::new()));
-        let values_ptr = Arc::clone(&values);
-
-        // SOLVE FOR THE INDUCTION VARIABLES, AND SUBSTITUTE INTO THE REWRITTEN RETURN EXPRESSION
-        // ======================================================================================
-
-        tracing::debug!(model=%generator_model,comprehension=%self,"Minion solving comprehnesion (ac mode)");
-
-        minion.solve(Box::new(move |sols| {
-            // TODO: deal with represented names if induction variables are abslits.
-            let values = &mut *values_ptr.lock().unwrap();
-            values.push(sols);
-            true
-        }))?;
-
-        let values = values.lock().unwrap().clone();
-
-        let mut return_expressions = vec![];
-
-        for value in values {
-            // convert back to an expression
-
-            let return_expression_submodel = return_expression_model.as_submodel().clone();
-            let child_symtab = return_expression_submodel.symbols().clone();
-            let return_expression = return_expression_submodel.into_single_expression();
-
-            // we only want to substitute induction variables.
-            // (definitely not machine names, as they mean something different in this scope!)
-            let value: HashMap<_, _> = value
-                .into_iter()
-                .filter(|(n, _)| self.induction_vars.contains(n))
-                .collect();
-
-            let value_ptr = Arc::new(value);
-            let value_ptr_2 = Arc::clone(&value_ptr);
-
-            // substitute in the values for the induction variables
-            let return_expression = return_expression.transform_bi(&move |x: Atom| {
-                let Atom::Reference(ref ptr) = x else {
-                    return x;
-                };
-
-                // is this referencing an induction var?
-                let Some(lit) = value_ptr_2.get(&ptr.name()) else {
-                    return x;
-                };
-
-                Atom::Literal(lit.clone())
-            });
-
-            // Copy the return expression's symbols into parent scope.
-
-            // For variables in the return expression with machine names, create new declarations
-            // for them in the parent symbol table, so that the machine names used are unique.
-            //
-            // Store the declaration translations in `machine_name_translations`.
-            // These are stored as a map of (old declaration id) -> (new declaration ptr), as
-            // declaration pointers do not implement hash.
-            //
-            let mut machine_name_translations: HashMap<ObjId, DeclarationPtr> = HashMap::new();
-
-            // Populate `machine_name_translations`
-            for (name, decl) in child_symtab.into_iter_local() {
-                // do not add givens for induction vars to the parent symbol table.
-                if value_ptr.get(&name).is_some()
-                    && matches!(&decl.kind() as &DeclarationKind, DeclarationKind::Given(_))
-                {
-                    continue;
-                }
-
-                let Name::Machine(_) = &name else {
-                    bug!(
-                        "the symbol table of the return expression of a comprehension should only contain machine names"
-                    );
-                };
-
-                let id = decl.id();
-                let new_decl = symtab.gensym(&decl.domain().unwrap());
-
-                machine_name_translations.insert(id, new_decl);
-            }
-
-            // Update references to use the new delcarations.
-            #[allow(clippy::arc_with_non_send_sync)]
-            let return_expression = return_expression.transform_bi(&move |atom: Atom| {
-                if let Atom::Reference(ref decl) = atom
-                    && let id = decl.id()
-                    && let Some(new_decl) = machine_name_translations.get(&id)
-                {
-                    Atom::Reference(new_decl.clone())
-                } else {
-                    atom
-                }
-            });
-
-            return_expressions.push(return_expression);
-        }
-
-        Ok(return_expressions)
     }
 
     pub fn return_expression(self) -> Expression {
