+++ crates/conjure-cp-rules/src/lib.rs
@@ -10,8 +10,8 @@ mod base;
 mod bottom_up_adaptor;
 mod bubble;
 mod cnf;
+mod comprehensions;
 mod constant_eval;
-mod expand_comprehension;
 mod matrix;
 mod minion;
 mod normalisers;
